% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdataz.R
\name{exdataz}
\alias{exdataz}
\title{Do interpolation for original time measurements and covariates.}
\usage{
exdataz(x, z, p, id, idmat, n = round(365 * diff(range(x))))
}
\arguments{
\item{x}{vector of ages (assume x is centered).}

\item{z}{data.frame of time independent covariates (assume z is centered).}

\item{p}{number of columns in data.frame z.}

\item{id}{factor of subject identifiers.}

\item{idmat}{matrix of unique id, note that this matrix has been setted to 1 column.}

\item{n}{time (age) measures after extension for computing aphv (default is round(365*diff(range(x)))).}
}
\value{
a data frame including expanded x(age), centralization covariates z and corresponding id.
}
\description{
This function evenly partitions time measurements and produces additional ones for an 
individual in his/her age range. It then uses \code{\link{predict.gmusim}} to obtain additional fitted 
values to calculate age at peak velocity (apv), peak velocity (pv) and height (or weight) at peak velocity (ypv).
}
\details{
For some individuals, the number of measurements is small. In order to calculate accurate apv 
         (age at peak velocity), pv (peak velocity) and ypv (height at peak velocity or weight at peak 
         velocity), it is necessary to perform interpolation for original time measurements and obtain 
         additional predictions for each individual. This function extends time (age) measurements 
         between min(time) and max(time), as well as covariates z, note that output of z has been centralized.
}
\examples{
require(sitar)
data(heights)
x <- heights$age
h <- heights$height
id <- heights$id
men <- heights$men
men <- abs(men)-mean(abs(men))
z <- data.frame(z1=men)
p <- 1
idmat <- matrix(unique(id), ncol = 1)
newdata <- exdataz(x, z, p, id, idmat)
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}, Man-Yu Wong \email{mamywong@ust.hk}
}
