% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lim.R
\name{lim}
\alias{lim}
\title{Linear interpolation method to calculate y value at time t0.}
\usage{
lim(x, y, id, t0)
}
\arguments{
\item{x}{vector of ages.}

\item{y}{vector of measurements.}

\item{id}{factor of subject identifiers.}

\item{t0}{target age value.}
}
\value{
a data frame including id and corresponding y value at t0.
}
\description{
This function uses linear interpolation method to calculate corresponding y value at time t0.
}
\details{
For each individual's measurements in growth data set, the first time measurement is usually not the same,
         people are interested in some measurement index of certain age, for example height of age 6 years old 
         for each id. This function can help us compute corresponding measurement at interested age.
}
\examples{
require(sitar)
data(heights)
x <- heights$age
h <- heights$height
id <- heights$id
## height at age 9
h9 <- lim(x,h,id,9)  
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk}, Man-Yu Wong \email{mamywong@ust.hk}
}
